Side Bar Navigation works, clean up code
