+++ src/components/Navigation/index.js
@@ -63,57 +63,4 @@ const NavigationNonAuth = () => (
   </ul>
 );
 
-export const mainListItems = (
-  <div>
-    <ListItem button>
-      <ListItemIcon>
-        <DashboardIcon />
-      </ListItemIcon>
-      <ListItemText primary="Dashboard" />
-    </ListItem>
-    <ListItem button>
-      <ListItemIcon>
-        <ViewListIcon />
-      </ListItemIcon>
-      <ListItemText primary="Accounts" />
-    </ListItem>
-    <ListItem button>
-      <ListItemIcon>
-        <TimelapseIcon />
-      </ListItemIcon>
-      <ListItemText primary="Recent Transactions" />
-    </ListItem>
-    <ListItem button>
-      <ListItemIcon>
-        <FindInPageIcon />
-      </ListItemIcon>
-      <ListItemText primary="Research" />
-    </ListItem>
-    <ListItem button>
-      <ListItemIcon>
-        <HelpIcon />
-      </ListItemIcon>
-      <ListItemText primary="Help" />
-    </ListItem>
-  </div>
-);
-
-export const secondaryListItems = (
-  <div>
-    <ListSubheader inset>My Account</ListSubheader>
-    <ListItem button>
-      <ListItemIcon>
-        <AccountBoxIcon />
-      </ListItemIcon>
-      <ListItemText primary="Account Details" />
-    </ListItem>
-    <ListItem button>
-      <ListItemIcon>
-        <SettingsIcon />
-      </ListItemIcon>
-      <ListItemText primary="Settings" />
-    </ListItem>
-  </div>
-);
-
 export default Navigation;
